+++ NEWS.md
@@ -1,4 +1,4 @@
-# rdflib 0.1.0
+# rdflib 0.1.0 (2018-03-02)
 
 ## New Features